src/orchestrator/pyproject.toml
[project]
name = "eval-microservice"
version = "0.1.0"
description = "Text completeness and image quality evaluation microservice (FastAPI, Vertex AI, heuristics)"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "orjson>=3.10",
  "python-dotenv>=1.0",
  "structlog>=24.1",
  "google-cloud-aiplatform>=1.64",
  "google-cloud-storage>=2.16",
  "opencv-python-headless>=4.10",
  "Pillow>=10.4",
  "numpy>=1.26",
  "black>=24.1",
  "flake8>=7.0.0",
  "pre-commit>=3.6.0",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "httpx>=0.27",
  "google-adk"
]
